requirements.txt
customtkinter==5.2.2
numpy==2.0.2
pillow==11.0.0
PyAudio==0.2.14
pydub==0.25.1
ru_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.8.0/ru_core_news_sm-3.8.0-py3-none-any.whl#sha256=69978d47b43e2c4f329bebdb155e8e9d3861bba1a58ba25551419dae7d7e07fc
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.12.0
